pyproject.toml
[tool.poetry]
name = "atis-intent-classifier"
version = "0.1.0"
description = "atis intent classifier"
authors = ["vasilis <vsyropou5@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10, <3.12"
numpy = "^1.26.2"
pandas = "^2.1.3"
scikit-learn = "^1.3.2"
python-dotenv = "^1.0.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
cloudpickle = "^3.0.0"
keras = "^2.15.0"
tensorflow = "^2.15.0"
pydantic = "^2.5.2"
matplotlib = "^3.8.2"
nltk = "^3.8.1"
httpx = "^0.25.2"
tensorflow-hub = "^0.15.0"
tf-sentencepiece = "^0.1.90"
tensorboard = "^2.15.1"
tensorflow-text = "^2.15.0"
lightgbm = "^4.1.0"
seaborn = "^0.13.0"
numba = "^0.58.1"
datasets = "^2.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-benchmark = "^4.0.0"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
